frontend: make flyout resizable and introduce a global app context state
