+++ src/packages/frontend/_projects.sass
@@ -131,3 +131,6 @@ li.cc-project-home-recent-files
   cursor: pointer
   &:hover
     background-color: $COL_GRAY_LLL
+
+div.cc-project-flyout-dragbar:hover
+  background-color: $COL_GRAY_L
