+++ src/packages/frontend/components/icon.tsx
@@ -18,11 +18,11 @@ import {
   ApiOutlined,
   AppstoreOutlined,
   AreaChartOutlined,
-  ArrowsAltOutlined,
   ArrowDownOutlined,
   ArrowLeftOutlined,
   ArrowRightOutlined,
   ArrowUpOutlined,
+  ArrowsAltOutlined,
   AudioOutlined,
   BackwardOutlined,
   BellFilled,
@@ -121,6 +121,7 @@ import {
   MailOutlined,
   MedicineBoxOutlined,
   MehOutlined,
+  MenuOutlined,
   MinusCircleOutlined,
   MinusOutlined,
   MinusSquareOutlined,
@@ -187,8 +188,8 @@ import {
   UserDeleteOutlined,
   UserOutlined,
   UsergroupAddOutlined,
-  VerticalRightOutlined,
   VerticalAlignMiddleOutlined,
+  VerticalRightOutlined,
   VideoCameraOutlined,
   WalletOutlined,
   WarningOutlined,
@@ -393,6 +394,7 @@ const IconSpec = {
   markdown: { IconFont: "markdown" },
   mask: { IconFont: "mask" },
   medkit: MedicineBoxOutlined,
+  "menu-outlined": MenuOutlined,
   meh: MehOutlined,
   microchip: { IconFont: "microchip" },
   "minus-circle": MinusCircleOutlined,
