+++ src/packages/frontend/index.sass
@@ -429,6 +429,10 @@ div.cc-project-fixedtab:hover
 span.cc-project-fixedtab-close:hover
   background-color: $COL_GRAY_L0
 
+span.cc-project-fixedtab-drag:hover
+  background-color: $COL_GRAY_L0
+  cusor: move
+
 div.cc-project-flyout-file-item:hover
   background-color: $COL_GRAY_L0
